src/vhost.c: enable client brick creation.

+++ src/vhost.c
@@ -84,6 +84,7 @@ struct pg_vhost_state {
 	struct pg_vhost_socket *socket;
 	rte_atomic32_t allow_queuing;
 	int vid;
+	uint64_t flags;
 	struct rte_mbuf *in[PG_MAX_PKTS_BURST];
 	struct rte_mbuf *out[PG_MAX_PKTS_BURST];
 	PG_PKTS_COUNT_TYPE tx_bytes; /* TX: [vhost] --> VM */
@@ -94,12 +95,12 @@ struct pg_vhost_state {
 #endif
 };
 
-static int new_vm(int dev);
-static void destroy_vm(int dev);
+static int on_new_device(int dev);
+static void on_destroy_device(int dev);
 
 static const struct vhost_device_ops virtio_net_device_ops = {
-	.new_device = new_vm,
-	.destroy_device = destroy_vm,
+	.new_device = on_new_device,
+	.destroy_device = on_destroy_device,
 };
 
 /* head of the socket list */
@@ -137,11 +138,13 @@ static int vhost_burst(struct pg_brick *brick, enum pg_side from,
 	uint64_t tx_bytes = 0;
 	const int check_atomic = VHOST_GET_CHECK_ATOMIC(state);
 
+	printf("Trying lock...%d %d\n", !rte_atomic32_test_and_set(
+		&state->allow_queuing), check_atomic);
 	/* Try lock */
 	if (unlikely(check_atomic &&
 		     !rte_atomic32_test_and_set(&state->allow_queuing)))
 		return 0;
-
+	printf("Lock ok\n");
 	virtio_net = state->vid;
 	pkts_count = pg_packets_pack(state->out, pkts, pkts_mask);
 	if (pkts_count > 32) {
@@ -180,6 +183,17 @@ static int vhost_burst(struct pg_brick *brick, enum pg_side from,
 
 #define TCP_PROTOCOL_NUMBER 6
 #define UDP_PROTOCOL_NUMBER 17
+
+/* To know if a vhost brick is server or not. */
+static int pg_vhost_is_server(struct pg_brick *brick)
+{
+	struct pg_vhost_state *state =
+		pg_brick_get_state(brick, struct pg_vhost_state);
+
+	if(state->flags & 1)
+		return 0;
+	return 1;
+}
 
 #ifdef PG_VHOST_FASTER_YET_BROKEN_POLL
 
@@ -284,8 +298,15 @@ static void vhost_create_socket(struct pg_vhost_state *state, uint64_t flags,
 	int ret;
 
 	path = g_strdup_printf("%s/qemu-%s", sockets_path, state->brick.name);
-	g_remove(path);
-
+	/* If the socket is CLIENT do NOT destroy the socket. */
+	if((flags & 1) == 0)
+	{
+		g_remove(path);
+	}
+	else
+	{
+		flags = flags | 31;
+	}
 	printf("New vhost-user socket: %s, zero-copy %s\n", path,
 	       (flags & RTE_VHOST_USER_DEQUEUE_ZERO_COPY) ?
 	       "enable" : "disable");
@@ -315,7 +336,6 @@ static void vhost_create_socket(struct pg_vhost_state *state, uint64_t flags,
 			path);
 		goto free_exit;
 	}
-
 	pthread_mutex_lock(&mutex);
 
 	s = g_new0(struct pg_vhost_socket, 1);
@@ -350,6 +370,7 @@ static int vhost_init(struct pg_brick *brick, struct pg_brick_config *config,
 	vhost_config = (struct pg_vhost_config *) config->brick_config;
 	state->output = vhost_config->output;
 	state->vid = -1;
+	state->flags = vhost_config->flags;
 	PG_PKTS_COUNT_SET(state->rx_bytes, 0);
 	PG_PKTS_COUNT_SET(state->tx_bytes, 0);
 	pg_vhost_enable(brick, enable_freacture_mask);
@@ -392,20 +413,19 @@ struct pg_brick *pg_vhost_new(const char *name, uint64_t flags,
 	return ret;
 }
 
-
 static void vhost_destroy(struct pg_brick *brick, struct pg_error **errp)
 {
-	struct pg_vhost_state *state;
-
-	state = pg_brick_get_state(brick, struct pg_vhost_state);
-	rte_vhost_driver_unregister(state->socket->path);
+	struct pg_vhost_state *state =
+		pg_brick_get_state(brick, struct pg_vhost_state);
 
 	pthread_mutex_lock(&mutex);
-	g_remove(state->socket->path);
-	LIST_REMOVE(state->socket, socket_list);
+	rte_vhost_driver_unregister(state->socket->path);
 	g_free(state->socket->path);
 	g_free(state->socket);
-
+	LIST_REMOVE(state->socket, socket_list);
+	/* If the socket is client, do NOT destroy the existing socket. */
+	if(pg_vhost_is_server(brick))
+		g_remove(state->socket->path);
 	pthread_mutex_unlock(&mutex);
 }
 
@@ -431,8 +451,9 @@ static uint64_t tx_bytes(struct pg_brick *brick)
 		pg_brick_get_state(brick, struct pg_vhost_state)->tx_bytes);
 }
 
-static int new_vm(int dev)
+static int on_new_device(int dev)
 {
+	printf("---------------I'm in the callback...-------\n");
 	struct pg_vhost_socket *s = NULL;
 	char buf[256];
 
@@ -454,7 +475,7 @@ static int new_vm(int dev)
 	return 0;
 }
 
-static void destroy_vm(int dev)
+static void on_destroy_device(int dev)
 {
 	struct pg_vhost_socket *s = NULL;
 	char buf[256];
